Implemented Id validations

+++ src/main/java/lk/ijse/gdse67/springposbackend/controller/CustomerController.java
@@ -1,10 +1,12 @@
 package lk.ijse.gdse67.springposbackend.controller;
 
+import lk.ijse.gdse67.springposbackend.customStatusCodes.SelectedCustomerStatus;
 import lk.ijse.gdse67.springposbackend.dto.CustomerStatus;
 import lk.ijse.gdse67.springposbackend.dto.impl.CustomerDto;
 import lk.ijse.gdse67.springposbackend.exception.CustomerNotFoundException;
 import lk.ijse.gdse67.springposbackend.exception.DataPersistException;
 import lk.ijse.gdse67.springposbackend.service.CustomerService;
+import lk.ijse.gdse67.springposbackend.util.Regex;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
@@ -41,14 +43,24 @@ public class CustomerController {
 
     @GetMapping(value = "/{propertyId}", produces = MediaType.APPLICATION_JSON_VALUE)
     public CustomerStatus getCustomer(@PathVariable("propertyId") String propertyId) {
-        return customerService.getCustomer(propertyId);
+        boolean isCustomerIdValid = Regex.CUSTOMER_ID.validate(propertyId);
+        if (isCustomerIdValid) {
+            return customerService.getCustomer(propertyId);
+        }else{
+            return new SelectedCustomerStatus(1, "Customer Id Invalid");
+        }
     }
 
     @PutMapping(value = "/{propertyId}",consumes = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<Void> updateCustomer(@PathVariable("propertyId") String propertyId, @RequestBody CustomerDto customerDto) {
+        boolean isCustomerIdValid = Regex.CUSTOMER_ID.validate(propertyId);
         try {
-            customerService.updateCustomer(propertyId, customerDto);
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            if(isCustomerIdValid){
+                customerService.updateCustomer(propertyId, customerDto);
+                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            }else{
+                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            }
         }catch (CustomerNotFoundException | DataPersistException e){
             e.printStackTrace();
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
@@ -60,9 +72,14 @@ public class CustomerController {
 
     @DeleteMapping(value = "/{propertyId}")
     public ResponseEntity<Void> deleteCustomer(@PathVariable("propertyId") String propertyId) {
+        boolean isCustomerIdValid = Regex.CUSTOMER_ID.validate(propertyId);
         try {
-            customerService.deleteCustomer(propertyId);
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            if(isCustomerIdValid){
+                customerService.deleteCustomer(propertyId);
+                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            }else{
+                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            }
         }catch (CustomerNotFoundException e){
             e.printStackTrace();
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
