+++ src/main/java/lk/ijse/gdse67/springposbackend/controller/ItemController.java
@@ -6,6 +6,7 @@ import lk.ijse.gdse67.springposbackend.dto.impl.ItemDto;
 import lk.ijse.gdse67.springposbackend.exception.DataPersistException;
 import lk.ijse.gdse67.springposbackend.exception.ItemNotFoundException;
 import lk.ijse.gdse67.springposbackend.service.ItemService;
+import lk.ijse.gdse67.springposbackend.util.Regex;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
@@ -42,14 +43,24 @@ public class ItemController {
 
     @GetMapping(value = "/{propertyId}", produces = MediaType.APPLICATION_JSON_VALUE)
     public ItemStatus getItem(@PathVariable("propertyId") String propertyId){
-        return itemService.getItem(propertyId);
+        boolean isItemIdValid = Regex.ITEM_ID.validate(propertyId);
+        if (isItemIdValid){
+            return itemService.getItem(propertyId);
+        }else{
+            return new SelectedItemCodes(1, "Item Id Invalid");
+        }
     }
 
     @PutMapping(value = "/{propertyId}", produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<Void> updateItem(@PathVariable("propertyId") String propertyId, @RequestBody ItemDto itemDto){
+        boolean isItemIdValid = Regex.ITEM_ID.validate(propertyId);
         try{
-            itemService.updateItem(propertyId, itemDto);
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            if(isItemIdValid){
+                itemService.updateItem(propertyId, itemDto);
+                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            }else{
+                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            }
         }catch (ItemNotFoundException e){
             e.printStackTrace();
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
@@ -61,9 +72,14 @@ public class ItemController {
 
     @DeleteMapping(value = "/{propertyId}")
     public ResponseEntity<Void> deleteItem(@PathVariable("propertyId") String propertyId){
+        boolean isItemIdValid = Regex.ITEM_ID.validate(propertyId);
         try{
-            itemService.deleteItem(propertyId);
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            if(isItemIdValid){
+                itemService.deleteItem(propertyId);
+                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            }else{
+                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            }
         }catch (ItemNotFoundException e){
             e.printStackTrace();
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
